# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Microservicio en Go para consultar los catalogos legados de creditos hipotecarios.",
        "title": "API de Catalogos de Creditos Hipotecarios",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8082",
    "basePath": "/v1/bfcl/mortgage-loan",
    "paths": {
        "/catalogs/{catalog}": {
            "post": {
                "description": "Retorna los elementos de un catálogo dado (simulando comportamiento legado).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "catalogs"
                ],
                "summary": "Obtener Catálogo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Nombre del catálogo (ej. Destino, Comunas, SegurosIncendio)",
                        "name": "catalog",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Canal (ej. WEB)",
                        "name": "X-Channel",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Comercio (ej. FALABELLA)",
                        "name": "X-Commerce",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID de transacción (ej. 12345)",
                        "name": "X-Transaction-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Entorno a inyectar: dummy, real, o java (por defecto: dummy)",
                        "name": "X-Backend-Env",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handler.catalogItemDTO"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponseDTO"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponseDTO"
                        }
                    },
                    "402": {
                        "description": "Payment Required",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponseDTO"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.internalServerErrorDTO"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handler.catalogItemDTO": {
            "type": "object",
            "properties": {
                "codigo_adm": {
                    "type": "string"
                },
                "descripcion": {
                    "type": "string"
                }
            }
        },
        "handler.errorResponseDTO": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "detail": {
                    "type": "string"
                },
                "errors_detail": {
                    "type": "string"
                },
                "messages": {
                    "type": "array",
                    "items": {}
                }
            }
        },
        "handler.internalServerErrorDTO": {
            "type": "object",
            "properties": {
                "mensaje": {
                    "type": "string"
                },
                "statusCode": {
                    "type": "integer"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        }
    }
}